// src/types/organizationDetailsTypes.ts
export type AddressType = {
  streetName: string;
  streetNumber: string;
  postalCode: string;
  city: string;
  country: string;
};

export type ContactDetailType = {
  emailAddress: string;
  telephone: string;
  website: string;
};

export type OrganizationDetailsType = {
  migrationMode: boolean;
  code: string;
  description: string;
  bankAccount: string;
  vatAccountNumber: string;
  companyAccountNumber: string;
  contactDetails: ContactDetailType;
  address: AddressType;
};